Add a tentative performance fix
The patch is based on Uri's comment here:

https://canva.slack.com/archives/C7883Q2JH/p1651135567415089

It intends to replace commits 86e646a1 and 1c6073a9 from the previous
version. We don't have a good understanding of why it works better
but empirically it allows to bring the sync duration down from 400 sec
to 75 sec in the same environment.

It has to be extensively tested before being deployed wider.